refactor(tx): drop stamps and use plain functions
