+++ src/ae/aepp.js
@@ -25,7 +25,6 @@
 import Ae from './'
 import Aens from './aens'
 import AeppRpc from '../utils/aepp-wallet-communication/rpc/aepp-rpc'
-import Tx from '../tx/tx'
 import Oracle from './oracle'
 import asyncInit from '../utils/async-init'
 
@@ -41,4 +40,4 @@ import asyncInit from '../utils/async-init'
  * @param {Object} [options={}] - Initializer object
  * @return {Object} Aepp instance
  */
-export default Ae.compose(asyncInit, Tx, Oracle, Aens, AeppRpc)
+export default Ae.compose(asyncInit, Oracle, Aens, AeppRpc)
