+++ src/index.js
@@ -31,7 +31,7 @@ export { default as Ae } from './ae'
 export { default as Node } from './node'
 export { default as NodePool } from './node-pool'
 export { default as Tx } from './tx'
-export { default as Transaction } from './tx/tx'
+export { default as Transaction } from './tx'
 export { default as verifyTransaction } from './tx/validator'
 export { default as AccountBase } from './account/base'
 export { default as AccountMultiple } from './account/multiple'
